requirements.txt
numpy~=1.26.4
docplex~=2.27.239
opencv-python~=4.9.0.80
matplotlib~=3.8.4
pandas~=2.2.2
scikit-image~=0.22.0
scikit-learn~=1.5.0
mediapipe~=0.10.11
tensorflow~=2.16.1
pillow~=10.3.0
torch~=2.3.0
pygame~=2.5.2
joblib~=1.4.2
tqdm~=4.66.4
seaborn~=0.13.2